fix initialization in splashScreen
